Move config.py into sc_flask module.
Issue #3564
